% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_labkey.R
\name{connect.to.labkey}
\alias{connect.to.labkey}
\title{Create connection to Labkey}
\usage{
connect.to.labkey(file = "~/.labkey.cred")
}
\arguments{
\item{file}{labkey credential file path, default is ~/.labkey.cred}
}
\description{
Create connection to Labkey
}
