% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_labkey.R
\docType{data}
\name{crf.prefixes}
\alias{crf.prefixes}
\title{Tibble for clinical table prefixes current 2019 Oct}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.}
\usage{
crf.prefixes
}
\description{
Tibble for clinical table prefixes current 2019 Oct
}
\keyword{datasets}
