% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oc_data.R
\name{download.all.tables}
\alias{download.all.tables}
\title{USE download.wide.format instead. This is the old method.}
\usage{
download.all.tables(ocs, occams_ids = NULL, missing = NULL,
  versions = "z1", verbose = T)
}
\arguments{
\item{occams_ids}{Array of occams identifiers, this is optional. The default returns all patients, better to use get.patients(...) if you want this.}

\item{missing}{Replace missing values with this string, default is NA}

\item{verbose}{}
}
\description{
USE download.wide.format instead. This is the old method.
}
