% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oc_data.R
\name{download.wide.format}
\alias{download.wide.format}
\title{Download in wide format the Labkey data for all or a selected group of patients.}
\usage{
download.wide.format(ocs, occams_ids = NULL, missing = NULL,
  versions = "z1", verbose = T)
}
\arguments{
\item{ocs}{Labkey connection object from connect.to.labkey()}

\item{occams_ids}{Array of occams identifiers, this is optional. The default returns all patients, better to use get.patients(...) if you want this.}

\item{missing}{Replace missing values with this string, default is NA}

\item{versions}{The table versions used from labkey, default is 'z1'}

\item{verbose}{}
}
\description{
Download in wide format the Labkey data for all or a selected group of patients.
}
\author{
skillcoyne
}
