% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oc_data.R
\name{get.patients}
\alias{get.patients}
\title{Get selected patients in the wide table format}
\usage{
get.patients(ocs, occams_ids, verbose = T)
}
\arguments{
\item{ocs}{Connection from connect.to.labkey()}

\item{occams_ids}{Array of occams identifiers to select}
}
\description{
Get selected patients in the wide table format
}
\author{

}
