% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_labkey.R
\name{get.prefixes}
\alias{get.prefixes}
\title{Get the table prefixes from the schema}
\usage{
get.prefixes(ocs, version = "z1")
}
\arguments{
\item{ocs}{Labkey connection}

\item{version}{version for tables, default 'z1'}
}
\description{
Get the table prefixes from the schema
}
\author{
skillcoyne
}
