% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_labkey.R
\name{get.tables.by.crf}
\alias{get.tables.by.crf}
\title{Get tables for the given CRF term (see crf.prefixes) or prefix.}
\usage{
get.tables.by.crf(ocs, crf)
}
\arguments{
\item{ocs}{Labkey connection}
}
\description{
Get tables for the given CRF term (see crf.prefixes) or prefix.
}
\author{
skillcoyne
}
