% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_labkey.R
\name{list.clinical.tables}
\alias{list.clinical.tables}
\title{List the clinical tables available in Labkey}
\usage{
list.clinical.tables(ocs, prefixes = NULL, versions = "z1")
}
\arguments{
\item{ocs}{Labkey connection}

\item{version}{version for tables, default 'z1'}
}
\description{
List the clinical tables available in Labkey
}
\author{
skillcoyne
}
