% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.demographics}
\alias{read.demographics}
\title{Read data from demographics CRF table.  
Strips out columns that include "Panther, CRF, age in months, and the "OpenClinica" informed consent columns.}
\usage{
read.demographics(ocs, table, occams_ids = NULL, rulesFile = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from demographics CRF table.  
Strips out columns that include "Panther, CRF, age in months, and the "OpenClinica" informed consent columns.
}
\author{

}
