% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.endpoint}
\alias{read.endpoint}
\title{Read data from endpoint CRF table.  
Strips out columns that include "OpenClinica(Start|Interview)Date}
\usage{
read.endpoint(ocs, tables, occams_ids = NULL, rulesFile = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from endpoint CRF table.  
Strips out columns that include "OpenClinica(Start|Interview)Date
}
\author{

}
