% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.exposures}
\alias{read.exposures}
\title{Read data from expoosures CRF tables.  
This includes: family history for OGC, family history of other cancers, and the general exposures tables
Recalculate BMI, create EX.CurrentBMI.c and EX.5YearsAgoBMI.c columns}
\usage{
read.exposures(ocs, tables, occams_ids = NULL, rulesFiles = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from expoosures CRF tables.  
This includes: family history for OGC, family history of other cancers, and the general exposures tables
Recalculate BMI, create EX.CurrentBMI.c and EX.5YearsAgoBMI.c columns
}
\author{

}
