% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.followup}
\alias{read.followup}
\title{Read data from followup, endpoint, and recurrence CRF tables.  
This requires a lot of cleaning around to get a single row per patient with the most recent follow up visit date, recurrence, and death. 
I have seen a few small inconsistencies in the dates, but they are days rather than years so for now this will work.  
TODO: add some date sanity checks}
\usage{
read.followup(ocs, tables, occams_ids = NULL, rulesFiles = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from followup, endpoint, and recurrence CRF tables.  
This requires a lot of cleaning around to get a single row per patient with the most recent follow up visit date, recurrence, and death. 
I have seen a few small inconsistencies in the dates, but they are days rather than years so for now this will work.  
TODO: add some date sanity checks
}
\author{

}
