% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.pathology}
\alias{read.pathology}
\title{Read data from pathology CRF table}
\usage{
read.pathology(ocs, table, occams_ids = NULL, rulesFile = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from pathology CRF table
}
\author{

}
