% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.prestage}
\alias{read.prestage}
\title{Read data from prestaging CRF tables
There are 6 tables that appear to relate only to metastatic pre-staging. In these there are multiple entries per-patient with no
clear rationale for merging. Currently not needed for any analysis so leaving these for people to deal with on their own if they want them.}
\usage{
read.prestage(ocs, tables, occams_ids = NULL, rulesFiles = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from prestaging CRF tables
There are 6 tables that appear to relate only to metastatic pre-staging. In these there are multiple entries per-patient with no
clear rationale for merging. Currently not needed for any analysis so leaving these for people to deal with on their own if they want them.
}
\author{

}
