% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.referral.diagnosis}
\alias{read.referral.diagnosis}
\title{Read data from referral & diagnosis CRF tables}
\usage{
read.referral.diagnosis(ocs, tables, occams_ids = NULL, rulesFile = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from referral & diagnosis CRF tables
}
\author{

}
