% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.surgery}
\alias{read.surgery}
\title{Read data from surgery CRF tables}
\usage{
read.surgery(ocs, table, occams_ids = NULL, rulesFile = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from surgery CRF tables
}
\author{

}
