% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.therapy}
\alias{read.therapy}
\title{Read data from therapy plan CRF tables
Merged all of the various new therapy tables. Now if you want to look at whether patients recieved neoadj or endoscopic therapy you need to select the 
columns 'TR.NeoAdj' or 'TR.Endo'.  Bit confusing but currently easiest.}
\usage{
read.therapy(ocs, tables, occams_ids = NULL, rulesFiles = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from therapy plan CRF tables
Merged all of the various new therapy tables. Now if you want to look at whether patients recieved neoadj or endoscopic therapy you need to select the 
columns 'TR.NeoAdj' or 'TR.Endo'.  Bit confusing but currently easiest.
}
\author{

}
