% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{read.treatment.plan}
\alias{read.treatment.plan}
\title{Read data from treatment plan CRF tables}
\usage{
read.treatment.plan(ocs, tables, occams_ids = NULL, rulesFiles = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Read data from treatment plan CRF tables
}
\author{

}
