% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oc_data.R
\name{download.all.tables}
\alias{download.all.tables}
\title{Download in wide format the Labkey data for all or a selected group of patients.}
\usage{
download.all.tables(ocs, occams_ids = NULL, missing = NULL,
  verbose = T)
}
\arguments{
\item{occams_ids}{Array of occams identifiers, this is optional. The default returns all patients, better to use get.patients(...) if you want this.}

\item{missing}{Replace missing values with this string, default is NA}

\item{verbose}{}
}
\description{
Download in wide format the Labkey data for all or a selected group of patients.
}
