% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_labkey.R
\name{get.patients}
\alias{get.patients}
\title{Get selected patients in the wide table format}
\usage{
get.patients(ocs, occams_ids, verbose = F)
}
\arguments{
\item{ocs}{Connection from connect.to.labkey()}

\item{occams_ids}{Array of occams identifiers to select}
}
\description{
Get selected patients in the wide table format
}
